Create non_linear_feistel_cipher.py
not working.
When saving the key to a string, since the first value is 0x3 = 0011, the first two zeroes are not considered, so the string of the key becomes "11...". To solve this, I added "{0:16b}" before the format function, to force the program to extend the string to 16 bits.
When printing the string, though, the first two bits are there but they are still empty, like this "  11....". Manually writing the binary value of the key seems to be a solution for now (commented line 70).

The real problem is the " 'int' object is not available " error, that happens every time I run the program (independently from how i initialize the key in the main function)