Task 8 done
Modified the main function as to use the other known plaintext/ciphertext pairs to check the quality of the keypairs that produced matchings.
In order for a keypair to survive the quality chack, it has to correctly encrypt/decrypt all 5 known (u,x) pairs, which are now  read from the input file.
Changed the generation of the random key lists in order to avoid repetitions (it's also a lot faster now) .
When generating a lot of random keys (close to the max possible amount of 2^16 = 65536, for example, N1 = N2 = N = 50000), the algorithm returns one pair that survives the quality check:
k1 = 0xccdf, k2 = 0x2452

If a lower amount of keys is used, it's very unlikely to generate both keys in their respective lists, so the program almost never returns any surviving keypairs.